requirements.txt
# ===== Core =====
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=41.0.0

# ===== Database =====
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ===== Data / Excel =====
pandas>=2.1.0
openpyxl>=3.1.0

# ===== HTTP / API =====
httpx>=0.25.0
requests>=2.31.0

# ===== Dashboard =====
streamlit>=1.28.0
streamlit-aggrid>=1.0.0
streamlit-autorefresh>=1.0.0
plotly>=5.18.0
altair>=5.1.0

# ===== Web Scraping (HTML 파싱 — 클라우드에서도 필요) =====
beautifulsoup4>=4.12.0
lxml>=4.9.0
# playwright는 로컬 전용 (requirements-local.txt 참조)

# ===== Analysis =====
scipy>=1.11.0
scikit-learn>=1.3.0
numpy>=1.26.0

# ===== Utilities =====
python-dateutil>=2.8.0
pillow>=10.1.0
loguru>=0.7.0
thefuzz>=0.20.0
